consolidate Role and TeamType enums into Member and Team model files
