+++ server/models/team.js
@@ -1,12 +1,34 @@
 import mongoose from "mongoose";
-import TeamType from "./teamType";
+
+const { Schema, SchemaTypes } = mongoose;
+const { ObjectId } = SchemaTypes;
+
+/**
+ * The different types of teams for Bits of Good.
+ * @enum {string}
+ */
+const TeamType = {
+  EXEC: "exec", // The executive board
+  PROJECT: "project", // Project teams
+  COMMITTEE: "committee" // Committees
+};
+
+/**
+ * Returns whether a string is a valid team type.
+ * @param {string} string A string
+ */
+TeamType.isType = function isType(string) {
+  return Object.values(TeamType).indexOf(string) !== -1;
+};
+
+// prevent modification of the TeamType enum
+Object.freeze(TeamType);
+
+export default TeamType;
 
 /**
  * The Team model for Bits of Good teams.
  */
-
-const { Schema, SchemaTypes } = mongoose;
-const { ObjectId } = SchemaTypes;
 
 const teamSchema = new Schema({
   // the name of the team in Bits of Good
@@ -38,4 +60,6 @@ const teamSchema = new Schema({
   }
 });
 
-export default (mongoose.models.Team || mongoose.model("Team", teamSchema));
+const Team = mongoose.models.Team || mongoose.model("Team", teamSchema);
+
+export { TeamType, Team };
