+++ test/event.model.test.js
@@ -1,6 +1,6 @@
 import * as chai from "chai";
 import Event from "../server/models/event";
-import Member from "../server/models/member";
+import { Member } from "../server/models/member";
 import { connection, clearModels } from "./util";
 
 /**
